requirements.txt
asgiref==3.7.2
attrs==23.2.0
certifi==2024.2.2
Django==4.1.13
djongo==1.3.6
dnspython==2.6.1
exceptiongroup==1.2.0
h11==0.14.0
idna==3.6
outcome==1.3.0.post0
packaging==23.2
pillow==10.2.0
pymongo==3.12.3
PySocks==1.7.1
pytesseract==0.3.10
pytz==2024.1
selenium==4.18.1
sniffio==1.3.1
sortedcontainers==2.4.0
sqlparse==0.2.4
trio==0.24.0
trio-websocket==0.11.1
typing_extensions==4.10.0
urllib3==2.2.1
wsproto==1.2.0
